libbpf-tools: refactor, move public methods to trace_helpers.c
