+++ libbpf-tools/execsnoop.h
@@ -13,6 +13,7 @@
 #define LAST_ARG (FULL_MAX_ARGS_ARR - ARGSIZE)
 
 struct event {
+	char comm[TASK_COMM_LEN];
 	pid_t pid;
 	pid_t tgid;
 	pid_t ppid;
@@ -20,7 +21,6 @@ struct event {
 	int retval;
 	int args_count;
 	unsigned int args_size;
-	char comm[TASK_COMM_LEN];
 	char args[FULL_MAX_ARGS_ARR];
 };
